+++ foo_ui_columns/pch.h
@@ -63,6 +63,7 @@
 #include <wincodec.h>
 #include <strsafe.h>
 #include <strstream>
+#include <icm.h>
 #include <d2d1_3.h>
 #include <d3d11_1.h>
 #include <dxgi1_6.h>